refactor stage 1
